# docs/contracts/openapi.v1.json
{
  "openapi": "3.1.0",
...
  "info": {
    "title": "Knotic Sales Agent API",
    "version": "1.0.0",
    "description": "Version 1 browser and private realtime voice-worker contracts."
  },
  "servers": [
    {"url": "https://api.example.invalid", "description": "Environment-specific HTTPS origin"}
  ],
  "security": [{"cookieAuth": []}],
...
  "paths": {
...
    "/api/v1/auth/login": {
      "get": {
        "operationId": "startLogin",
        "tags": ["Authentication"],
        "summary": "Start OIDC Authorization Code with PKCE using server-held state",
        "security": [],
        "parameters": [{"$ref": "#/components/parameters/ReturnTo"}],
        "responses": {
          "302": {
            "description": "Redirect to the configured OIDC issuer",
            "headers": {
              "Location": {"$ref": "#/components/headers/ExternalLocation"},
              "Cache-Control": {"schema": {"type": "string", "const": "no-store"}},
              "X-Request-ID": {"$ref": "#/components/headers/RequestId"}
            }
          },
          "400": {"$ref": "#/components/responses/BadRequest"},
          "429": {"$ref": "#/components/responses/RateLimited"},
          "500": {"$ref": "#/components/responses/InternalError"},
          "503": {"$ref": "#/components/responses/DependencyUnavailable"}
        }
      }
    },
    "/api/v1/auth/callback": {
      "get": {
        "operationId": "completeLogin",
        "tags": ["Authentication"],
        "summary": "Validate the single-use OIDC callback and establish the application session",
        "security": [],
        "parameters": [
          {"$ref": "#/components/parameters/AuthorizationCode"},
          {"$ref": "#/components/parameters/OidcState"}
        ],
        "responses": {
          "303": {
            "description": "Application session established and user agent redirected to the validated relative path",
            "headers": {
              "Location": {"$ref": "#/components/headers/ApplicationLocation"},
              "Set-Cookie": {"$ref": "#/components/headers/SessionCookie"},
              "Cache-Control": {"schema": {"type": "string", "const": "no-store"}},
              "X-Request-ID": {"$ref": "#/components/headers/RequestId"}
            }
          },
          "400": {"$ref": "#/components/responses/BadRequest"},
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "429": {"$ref": "#/components/responses/RateLimited"},
          "500": {"$ref": "#/components/responses/InternalError"},
          "503": {"$ref": "#/components/responses/DependencyUnavailable"},
          "504": {"$ref": "#/components/responses/DependencyTimeout"}
        }
      }
    },
    "/api/v1/auth/session": {
      "get": {
        "operationId": "getAuthenticatedSession",
        "tags": ["Authentication"],
        "summary": "Read the safe authenticated actor and session expiry",
        "responses": {
          "200": {
            "description": "Current authenticated application session",
            "headers": {
              "Cache-Control": {"schema": {"type": "string", "const": "no-store"}},
              "X-Request-ID": {"$ref": "#/components/headers/RequestId"}
            },
            "content": {"application/json": {"schema": {"$ref": "#/components/schemas/AuthenticatedSession"}}}
          },
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "403": {"$ref": "#/components/responses/Forbidden"},
          "429": {"$ref": "#/components/responses/RateLimited"},
          "500": {"$ref": "#/components/responses/InternalError"}
        }
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "operationId": "logout",
        "tags": ["Authentication"],
        "summary": "Revoke the server-managed application session and expire its cookie",
        "parameters": [
          {"$ref": "#/components/parameters/CsrfToken"},
          {"$ref": "#/components/parameters/IdempotencyKey"}
        ],
        "requestBody": {
          "required": true,
          "content": {"application/json": {"schema": {"$ref": "#/components/schemas/LogoutRequest"}}}
        },
        "responses": {
          "200": {
            "description": "Local session revoked; repeating the same idempotent command returns this result",
            "headers": {
              "Set-Cookie": {"$ref": "#/components/headers/ExpiredSessionCookie"},
              "Cache-Control": {"schema": {"type": "string", "const": "no-store"}},
              "X-Request-ID": {"$ref": "#/components/headers/RequestId"}
            },
            "content": {"application/json": {"schema": {"$ref": "#/components/schemas/LogoutResult"}}}
          },
          "400": {"$ref": "#/components/responses/BadRequest"},
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "403": {"$ref": "#/components/responses/Forbidden"},
          "409": {"$ref": "#/components/responses/Conflict"},
          "415": {"$ref": "#/components/responses/UnsupportedMediaType"},
          "422": {"$ref": "#/components/responses/UnprocessableEntity"},
          "429": {"$ref": "#/components/responses/RateLimited"},
          "500": {"$ref": "#/components/responses/InternalError"}
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "knotic_session",
        "description": "HttpOnly, Secure, SameSite=Lax server-managed browser session."
      },
      "workloadAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "short-lived workload token",
        "description": "Private voice-worker workload identity; mTLS may replace bearer transport at deployment."
      }
    },
    "parameters": {
      "ReturnTo": {
        "name": "return_to",
        "in": "query",
        "required": false,
        "schema": {"type": "string", "pattern": "^/(?!/)[^\\r\\n]*$", "maxLength": 1024, "default": "/"},
        "description": "Allowlisted application-relative destination; absolute and protocol-relative values are rejected."
      },
      "AuthorizationCode": {
        "name": "code",
        "in": "query",
        "required": true,
        "schema": {"type": "string", "minLength": 8, "maxLength": 4096},
        "description": "Sensitive, single-use OIDC authorization code; never logged."
      },
      "OidcState": {
        "name": "state",
        "in": "query",
        "required": true,
        "schema": {"type": "string", "minLength": 32, "maxLength": 1024},
        "description": "Opaque, single-use server-bound OIDC state; never logged."
      },
...
      "CsrfToken": {
        "name": "X-CSRF-Token",
        "in": "header",
        "required": true,
        "schema": {"type": "string", "minLength": 32, "maxLength": 512},
        "description": "Token bound to the authenticated server-managed browser session."
      },
      "IdempotencyKey": {
        "name": "Idempotency-Key",
        "in": "header",
        "required": true,
        "schema": {"type": "string", "minLength": 16, "maxLength": 128, "pattern": "^[!-~]+$"}
      },
...
    },
    "headers": {
      "RequestId": {"description": "Safe request identifier", "schema": {"$ref": "#/components/schemas/Uuid"}},
...
      "ExternalLocation": {"description": "Configured HTTPS OIDC authorization endpoint", "schema": {"type": "string", "format": "uri", "pattern": "^https://"}},
      "ApplicationLocation": {"description": "Validated application-relative redirect destination", "schema": {"type": "string", "pattern": "^/(?!/)"}},
      "SessionCookie": {"description": "Server-generated knotic_session cookie with HttpOnly, Secure, SameSite=Lax, and Path=/api; the value must never be logged.", "schema": {"type": "string"}},
      "ExpiredSessionCookie": {"description": "Expired knotic_session cookie with Max-Age=0 and the same security attributes.", "schema": {"type": "string"}},
...
      "RetryAfter": {"description": "Whole seconds before retry or operation poll", "schema": {"type": "integer", "minimum": 1}},
      "RateLimitLimit": {"description": "Quota for the current window", "schema": {"type": "integer", "minimum": 1}},
      "RateLimitRemaining": {"description": "Remaining requests in the current window", "schema": {"type": "integer", "minimum": 0}},
      "RateLimitReset": {"description": "UTC epoch seconds when the quota resets", "schema": {"type": "integer", "minimum": 1}}
    },
    "responses": {
      "BadRequest": {"$ref": "#/components/responses/ErrorResponse400"},
      "Unauthorized": {"$ref": "#/components/responses/ErrorResponse401"},
      "Forbidden": {"$ref": "#/components/responses/ErrorResponse403"},
...
      "Conflict": {"$ref": "#/components/responses/ErrorResponse409"},
      "UnsupportedMediaType": {"$ref": "#/components/responses/ErrorResponse415"},
      "UnprocessableEntity": {"$ref": "#/components/responses/ErrorResponse422"},
      "RateLimited": {
        "description": "Caller quota exhausted",
        "headers": {
          "Retry-After": {"$ref": "#/components/headers/RetryAfter"},
          "RateLimit-Limit": {"$ref": "#/components/headers/RateLimitLimit"},
          "RateLimit-Remaining": {"$ref": "#/components/headers/RateLimitRemaining"},
          "RateLimit-Reset": {"$ref": "#/components/headers/RateLimitReset"},
          "X-Request-ID": {"$ref": "#/components/headers/RequestId"}
        },
        "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorEnvelope"}}}
      },
      "InternalError": {"$ref": "#/components/responses/ErrorResponse500"},
      "DependencyUnavailable": {
        "description": "Required dependency unavailable",
        "headers": {
          "Retry-After": {"$ref": "#/components/headers/RetryAfter"},
          "X-Request-ID": {"$ref": "#/components/headers/RequestId"}
        },
        "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorEnvelope"}}}
      },
      "DependencyTimeout": {
        "description": "Required dependency exceeded the bounded deadline",
        "headers": {"X-Request-ID": {"$ref": "#/components/headers/RequestId"}},
        "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorEnvelope"}}}
      },
      "ErrorResponse400": {"description": "Malformed request or cursor", "headers": {"X-Request-ID": {"$ref": "#/components/headers/RequestId"}}, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorEnvelope"}}}},
      "ErrorResponse401": {"description": "Authentication required or invalid", "headers": {"X-Request-ID": {"$ref": "#/components/headers/RequestId"}}, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorEnvelope"}}}},
      "ErrorResponse403": {"description": "Authorization, origin, CSRF, or policy denied", "headers": {"X-Request-ID": {"$ref": "#/components/headers/RequestId"}}, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorEnvelope"}}}},
...
      "ErrorResponse409": {"description": "Version, sequence, idempotency, or lifecycle conflict", "headers": {"X-Request-ID": {"$ref": "#/components/headers/RequestId"}}, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorEnvelope"}}}},
      "ErrorResponse415": {"description": "Request media type unsupported", "headers": {"X-Request-ID": {"$ref": "#/components/headers/RequestId"}}, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorEnvelope"}}}},
      "ErrorResponse422": {"description": "Request schema validation failed", "headers": {"X-Request-ID": {"$ref": "#/components/headers/RequestId"}}, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorEnvelope"}}}},
      "ErrorResponse500": {"description": "Safe internal error", "headers": {"X-Request-ID": {"$ref": "#/components/headers/RequestId"}}, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorEnvelope"}}}}
    },
    "schemas": {
      "Uuid": {"type": "string", "format": "uuid"},
      "Timestamp": {"type": "string", "format": "date-time"},
...
      "AuthenticatedSession": {
        "type": "object",
        "additionalProperties": false,
        "required": ["authenticated", "actor", "expires_at"],
        "properties": {
          "authenticated": {"type": "boolean", "const": true},
          "actor": {"$ref": "#/components/schemas/AuthenticatedActor"},
          "expires_at": {"$ref": "#/components/schemas/Timestamp"}
        }
      },
      "AuthenticatedActor": {
        "type": "object",
        "additionalProperties": false,
        "required": ["actor_id", "tenant_id", "display_name", "roles"],
        "properties": {
          "actor_id": {"$ref": "#/components/schemas/Uuid"},
          "tenant_id": {"$ref": "#/components/schemas/Uuid"},
          "display_name": {"type": "string", "minLength": 1, "maxLength": 200},
          "roles": {"type": "array", "items": {"type": "string", "enum": ["CUSTOMER", "SALES_REP", "SUPERVISOR", "ADMIN"]}, "minItems": 1, "uniqueItems": true}
        }
      },
      "LogoutRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": ["all_sessions"],
        "properties": {"all_sessions": {"type": "boolean", "description": "When true, revoke all application sessions for the authenticated actor when policy permits."}}
      },
      "LogoutResult": {
        "type": "object",
        "additionalProperties": false,
        "required": ["authenticated", "revoked_sessions"],
        "properties": {
          "authenticated": {"type": "boolean", "const": false},
          "revoked_sessions": {"type": "integer", "minimum": 1}
        }
      },
...
      "ErrorEnvelope": {
        "type": "object",
        "additionalProperties": false,
        "required": ["error"],
        "properties": {"error": {"$ref": "#/components/schemas/ErrorObject"}}
      },
      "ErrorObject": {
        "type": "object",
        "additionalProperties": false,
        "required": ["code", "category", "message", "retryable", "request_id", "correlation_id", "details"],
        "properties": {
          "code": {"$ref": "#/components/schemas/ErrorCode"},
          "category": {"type": "string", "enum": ["VALIDATION", "AUTHENTICATION", "AUTHORIZATION", "POLICY", "NOT_FOUND", "CONFLICT", "RATE_LIMIT", "DEPENDENCY", "INTERNAL"]},
          "message": {"type": "string", "minLength": 1, "maxLength": 500},
          "retryable": {"type": "boolean"},
          "request_id": {"$ref": "#/components/schemas/Uuid"},
          "correlation_id": {"oneOf": [{"$ref": "#/components/schemas/Uuid"}, {"type": "null"}]},
          "details": {"type": "array", "maxItems": 50, "items": {"$ref": "#/components/schemas/ErrorDetail"}}
        }
      },
      "ErrorCode": {"type": "string", "enum": ["MALFORMED_JSON", "INVALID_CURSOR", "INVALID_HEADER", "AUTHENTICATION_REQUIRED", "AUTHENTICATION_INVALID", "FORBIDDEN", "CSRF_FAILED", "ORIGIN_DENIED", "POLICY_DENIED", "RESOURCE_NOT_FOUND", "VERSION_CONFLICT", "SEQUENCE_CONFLICT", "IDEMPOTENCY_CONFLICT", "SESSION_NOT_ACCEPTING_TURNS", "UNSUPPORTED_MEDIA_TYPE", "VALIDATION_FAILED", "RATE_LIMITED", "INTERNAL_ERROR", "DEPENDENCY_UNAVAILABLE", "DEPENDENCY_TIMEOUT"]},
      "ErrorDetail": {
        "type": "object",
        "additionalProperties": false,
        "required": ["field", "issue"],
        "properties": {
          "field": {"type": ["string", "null"], "maxLength": 200},
          "issue": {"type": "string", "minLength": 1, "maxLength": 500}
        }
      }
    }
  },
...
}